refactor(database): 重构 ptc-object ORM 模块，拆分子包并消除重复代码
- Template Method：新建 AbstractDataMapper 基类，DataMapperImpl / TransactionalDataMapper 仅覆写差异部分
- Strategy：提取 DatabaseDialect 接口（MySQL/SQLite/PostgreSQL），消除 Container 子类建表重复
- SRP：新建 EntityMapper，将 ResultSet 读取与实例创建从 AnalyzedClass 中分离
- Facade+Composite：ContainerOperatorImpl 拆分为 DataExecutor / LinkTableHandler / CollectionTableHandler
- 分包：内部实现移至 orm / container / operator / mapper 子包，公开 API 保持不变
- 修复 database-ioc 对 DataMapperImpl / AnalyzedClass 的导入路径
