--- module/database/database-ptc-object/src/main/kotlin/taboolib/expansion/ContainerSQL.kt
@@ -1,167 +1,0 @@
-package taboolib.expansion
-
-import taboolib.module.database.*
-
-class ContainerSQL(
-    host: String,
-    port: Int,
-    user: String,
-    password: String,
-    database: String,
-    flags: List<String> = emptyList(),
-    clearFlags: Boolean = false,
-    ssl: String? = null,
-) : Container<SQL>(HostSQL(host, port.toString(), user, password, database).also {
-    if (clearFlags) {
-        it.flags.clear()
-    }
-    it.flags.addAll(flags)
-    if (ssl != null) {
-        it.flags -= "useSSL=false"
-        it.flags += "sslMode=$ssl"
-    }
-}) {
-
-    override fun createTableObject(type: AnalyzedClass, name: String): Table<*, *> {
-        return Table(name, host) {
-            // 只有在没有 @Id 字段时才自动添加 id 主键
-            if (!type.members.any { it.isPrimary }) {
-                add { id() }
-            }
-            type.members.forEach { member ->
-                // 跳过 @Ignore 成员
-                if (member.isIgnored) return@forEach
-                // 跳过容器类型成员（它们存储在子表中）
-                if (member.isCollection) return@forEach
-                // @LinkTable 成员：创建外键列而非展开关联类
-                if (member.isLinkTable) {
-                    val linkedClass = AnalyzedClass.of(member.linkTableClass!!)
-                    val linkedPrimary = linkedClass.primaryMember!!
-                    val fkColumnName = member.linkTableColumn!!
-                    add(fkColumnName) {
-                        when {
-                            linkedPrimary.hasColumnType -> {
-                                val colType = linkedPrimary.columnTypeSQL!!
-                                if (colType.isRequired) type(colType, linkedPrimary.length) else type(colType)
-                            }
-                            linkedPrimary.isIndexedEnum -> type(ColumnTypeSQL.BIGINT)
-                            linkedPrimary.isString || linkedPrimary.isEnum -> {
-                                if (linkedPrimary.length < 0) type(ColumnTypeSQL.LONGTEXT)
-                                else type(ColumnTypeSQL.VARCHAR, linkedPrimary.length)
-                            }
-                            linkedPrimary.isUUID -> type(ColumnTypeSQL.CHAR, 36)
-                            else -> type(linkedPrimary.type())
-                        }
-                    }
-                    return@forEach
-                }
-                when {
-                    // 自定义列类型
-                    member.hasColumnType -> add(member.name) {
-                        val colType = member.columnTypeSQL!!
-                        if (colType.isRequired) type(colType, member.length) { options(member) }
-                        else type(colType) { options(member) }
-                    }
-                    // IndexedEnum（数值存储）
-                    member.isIndexedEnum -> add(member.name) {
-                        type(ColumnTypeSQL.BIGINT) { options(member) }
-                    }
-                    // 字符串
-                    member.isString || member.isEnum -> add(member.name) {
-                        // length == -1 时使用longtext
-                        if (member.length < 0) {
-                            type(ColumnTypeSQL.LONGTEXT) { options(member) }
-                        } else {
-                            type(ColumnTypeSQL.VARCHAR, member.length) { options(member) }
-                        }
-                    }
-                    // UUID
-                    member.isUUID -> add(member.name) {
-                        type(ColumnTypeSQL.CHAR, 36) { options(member) }
-                    }
-                    // 字节数组
-                    member.isByteArray -> add(member.name) {
-                        type(ColumnTypeSQL.BLOB) { options(member) }
-                    }
-                    // 其他类型
-                    else -> add(member.name) {
-                        val customType = if (member.isFlattenedCollection) {
-                            CustomTypeFactory.getCustomTypeForCollection(member.returnType, member.collectionElementType!!)
-                        } else {
-                            CustomTypeFactory.getCustomTypeByClass(member.returnType)
-                        }
-                        if (customType == null) {
-                            type(member.type()) { options(member) }
-                        } else {
-                            type(customType.typeSQL, customType.length) { options(member) }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    override fun createCollectionTableObject(
-        parentType: AnalyzedClass,
-        parentTableName: String,
-        member: AnalyzedClassMember,
-        childTableName: String
-    ): Table<*, *> {
-        val primaryMember = parentType.primaryMember!!
-        return Table(childTableName, host) {
-            add { id() }
-            // FK 列：引用主表的 @Id
-            add("parent_${primaryMember.name}") {
-                when {
-                    primaryMember.hasColumnType -> {
-                        val colType = primaryMember.columnTypeSQL!!
-                        if (colType.isRequired) type(colType, primaryMember.length) else type(colType)
-                    }
-                    primaryMember.isIndexedEnum -> type(ColumnTypeSQL.BIGINT)
-                    primaryMember.isString || primaryMember.isEnum -> {
-                        if (primaryMember.length < 0) type(ColumnTypeSQL.LONGTEXT)
-                        else type(ColumnTypeSQL.VARCHAR, primaryMember.length)
-                    }
-                    primaryMember.isUUID -> type(ColumnTypeSQL.CHAR, 36)
-                    else -> type(primaryMember.type())
-                }
-            }
-            if (member.isMap) {
-                add("map_key") { type(ColumnTypeSQL.VARCHAR, 512) }
-                add("map_value") { type(ColumnTypeSQL.VARCHAR, 512) }
-            } else {
-                add("value") { type(ColumnTypeSQL.VARCHAR, 512) }
-                if (member.isList) {
-                    add("sort_order") { type(ColumnTypeSQL.INT) }
-                }
-            }
-        }
-    }
-
-    private fun ColumnSQL.options(member: AnalyzedClassMember) {
-        if (member.isKey || member.isPrimary) {
-            options(ColumnOptionSQL.KEY)
-        }
-        if (member.isUniqueKey) {
-            options(ColumnOptionSQL.UNIQUE_KEY)
-        }
-        if (member.isNotNull) {
-            options(ColumnOptionSQL.NOTNULL)
-        }
-    }
-
-    private fun AnalyzedClassMember.type(): ColumnTypeSQL {
-        return when {
-            isBoolean -> ColumnTypeSQL.BOOLEAN
-            isByte -> ColumnTypeSQL.TINYINT
-            isShort -> ColumnTypeSQL.SMALLINT
-            isInt -> ColumnTypeSQL.INT
-            isLong -> ColumnTypeSQL.BIGINT
-            isFloat -> ColumnTypeSQL.FLOAT
-            isDouble -> ColumnTypeSQL.DOUBLE
-            isChar -> ColumnTypeSQL.INT
-            isByteArray -> ColumnTypeSQL.BLOB
-            else -> error("Unsupported type: $name (${returnType})")
-        }
-    }
-}
