+++ module/database/database-ptc-object/src/main/kotlin/taboolib/expansion/container/ContainerSQL.kt
@@ -1,0 +1,26 @@
+package taboolib.expansion.container
+
+import taboolib.module.database.*
+
+class ContainerSQL(
+    host: String,
+    port: Int,
+    user: String,
+    password: String,
+    database: String,
+    flags: List<String> = emptyList(),
+    clearFlags: Boolean = false,
+    ssl: String? = null,
+) : Container<SQL>(HostSQL(host, port.toString(), user, password, database).also {
+    if (clearFlags) {
+        it.flags.clear()
+    }
+    it.flags.addAll(flags)
+    if (ssl != null) {
+        it.flags -= "useSSL=false"
+        it.flags += "sslMode=$ssl"
+    }
+}) {
+
+    override val dialect: DatabaseDialect = MySQLDialect
+}
