+++ module/database/database-ptc-object/src/main/kotlin/taboolib/expansion/container/ContainerSQLite.kt
@@ -1,0 +1,9 @@
+package taboolib.expansion.container
+
+import taboolib.module.database.*
+import java.io.File
+
+class ContainerSQLite(file: File) : Container<SQLite>(HostSQLite(file)) {
+
+    override val dialect: DatabaseDialect = SQLiteDialect
+}
