+++ module/database/database-ptc-object/src/main/kotlin/taboolib/expansion/mapper/AbstractDataMapper.kt
@@ -1,25 +1,38 @@
-package taboolib.expansion
+package taboolib.expansion.mapper
 
+import taboolib.expansion.ContainerOperator
+import taboolib.expansion.Cursor
+import taboolib.expansion.DataMapper
+import taboolib.expansion.L2Cache
+import taboolib.expansion.Page
+import taboolib.expansion.operator.ContainerOperatorImpl
+import taboolib.expansion.orm.AnalyzedClass
+import taboolib.expansion.orm.EntityMapper
 import taboolib.module.database.*
-import java.sql.Connection
 import java.sql.ResultSet
 
 /**
- * 事务内的 DataMapper 实现
+ * DataMapper 公共基类（Template Method 模式）
  *
- * 使用事务感知的 ContainerOperator，所有操作共享同一个数据库连接。
- * 不支持嵌套事务，调用 transaction() 会抛出异常。
- *
- * 事务内不使用缓存读取（保证事务一致性），但写入操作会失效外层 L2 缓存。
+ * 将 DataMapperImpl 和 TransactionalDataMapper 的公共 CRUD 实现合并到此基类，
+ * 子类通过覆写差异方法来实现各自的特定行为：
+ * - operator 来源（动态查找 vs 构造注入）
+ * - 读操作是否走缓存
+ * - 游标查询支持
+ * - 事务、join、close 行为
  */
-class TransactionalDataMapper<T>(
-    private val type: Class<T>,
-    private val operator: ContainerOperator,
-    private val cache: L2Cache?,
-    private val sharedConnection: Connection? = null
-) : DataMapper<T> {
+abstract class AbstractDataMapper<T> : DataMapper<T> {
 
-    private val analyzedClass by lazy { AnalyzedClass.of(type) }
+    protected abstract val type: Class<T>
+    protected abstract val operator: ContainerOperator
+    protected abstract val cache: L2Cache?
+
+    protected val analyzedClass by lazy { AnalyzedClass.of(type) }
+
+    // === 读操作分派（默认直接执行，子类可覆写为走缓存）===
+
+    protected open fun <R> executeRead(key: String, vararg args: Any?, query: () -> R): R = query()
+    protected open fun <R> executeBeanRead(key: String, vararg args: Any?, query: () -> R): R = query()
 
     // === 插入 ===
 
@@ -49,23 +62,42 @@ class TransactionalDataMapper<T>(
         return keys
     }
 
-    // === 查询（事务内不使用缓存，直接查库）===
+    // === 查询 ===
 
-    override fun findById(id: Any, filter: Filter.() -> Unit): T? = operator.findOne(type, id, filter)
-    override fun findAll(id: Any, filter: Filter.() -> Unit): List<T> = operator.find(type, id, filter)
-    override fun findOne(filter: Filter.() -> Unit): T? = operator.getOne(type, filter)
-    override fun findAll(filter: Filter.() -> Unit): List<T> = operator.get(type, filter)
+    override fun findById(id: Any, filter: Filter.() -> Unit): T? {
+        return executeBeanRead("findById", id, filter) { operator.findOne(type, id, filter) }
+    }
+
+    override fun findAll(id: Any, filter: Filter.() -> Unit): List<T> {
+        return executeBeanRead("findAllById", id, filter) { operator.find(type, id, filter) }
+    }
+
+    override fun findOne(filter: Filter.() -> Unit): T? {
+        return executeRead("findOne", filter) { operator.getOne(type, filter) }
+    }
+
+    override fun findAll(filter: Filter.() -> Unit): List<T> {
+        return executeRead("findAll", filter) { operator.get(type, filter) }
+    }
 
     override fun findByIds(ids: List<Any>): List<T> {
         if (ids.isEmpty()) return emptyList()
-        return operator.findByIds(type, ids)
+        return executeRead("findByIds", ids) { operator.findByIds(type, ids) }
     }
 
     // === 基于 @Key 的查询 ===
 
-    override fun findByKey(data: T): List<T> = operator.findByKey(type, data as Any)
-    override fun findOneByKey(data: T): T? = operator.findOneByKey(type, data as Any)
-    override fun existsByKey(data: T): Boolean = operator.hasByKey(type, data as Any)
+    override fun findByKey(data: T): List<T> {
+        return executeRead("findByKey", data) { operator.findByKey(type, data as Any) }
+    }
+
+    override fun findOneByKey(data: T): T? {
+        return executeRead("findOneByKey", data) { operator.findOneByKey(type, data as Any) }
+    }
+
+    override fun existsByKey(data: T): Boolean {
+        return executeRead("existsByKey", data) { operator.hasByKey(type, data as Any) }
+    }
 
     override fun deleteByKey(data: T) {
         operator.deleteByKey(data as Any)
@@ -74,7 +106,9 @@ class TransactionalDataMapper<T>(
 
     // === 基于自增行 ID 的操作 ===
 
-    override fun findByRowId(rowId: Long): T? = operator.findByRowId(type, rowId)
+    override fun findByRowId(rowId: Long): T? {
+        return executeBeanRead("findByRowId", rowId) { operator.findByRowId(type, rowId) }
+    }
 
     override fun deleteByRowId(rowId: Long) {
         operator.deleteByRowId(rowId)
@@ -84,45 +118,45 @@ class TransactionalDataMapper<T>(
     // === 排序 ===
 
     override fun sort(row: String, limit: Int, filter: Filter.() -> Unit): List<T> {
-        return operator.sort(type, row, limit, filter)
+        return executeRead("sort", row, limit, filter) { operator.sort(type, row, limit, filter) }
     }
 
     override fun sortDescending(row: String, limit: Int, filter: Filter.() -> Unit): List<T> {
-        return operator.sortDescending(type, row, limit, filter)
+        return executeRead("sortDescending", row, limit, filter) { operator.sortDescending(type, row, limit, filter) }
     }
 
     // === 分页 ===
 
     override fun findPage(page: Int, size: Int, filter: Filter.() -> Unit): Page<T> {
-        val content = operator.getPage(type, page, size, filter)
-        val total = operator.count(filter)
+        val content = executeRead("findPage", page, size, filter) { operator.getPage(type, page, size, filter) }
+        val total = executeRead("count", filter) { operator.count(filter) }
         return Page(content, page, size, total)
     }
 
     override fun sortPage(row: String, page: Int, size: Int, filter: Filter.() -> Unit): Page<T> {
-        val content = operator.sortPage(type, row, page, size, filter)
-        val total = operator.count(filter)
+        val content = executeRead("sortPage", row, page, size, filter) { operator.sortPage(type, row, page, size, filter) }
+        val total = executeRead("count", filter) { operator.count(filter) }
         return Page(content, page, size, total)
     }
 
     override fun sortDescendingPage(row: String, page: Int, size: Int, filter: Filter.() -> Unit): Page<T> {
-        val content = operator.sortDescendingPage(type, row, page, size, filter)
-        val total = operator.count(filter)
+        val content = executeRead("sortDescendingPage", row, page, size, filter) { operator.sortDescendingPage(type, row, page, size, filter) }
+        val total = executeRead("count", filter) { operator.count(filter) }
         return Page(content, page, size, total)
     }
 
-    // === 游标查询 ===
+    // === 游标查询（默认不支持，子类可覆写）===
 
     override fun selectCursor(filter: Filter.() -> Unit): Cursor<T> {
-        return operator.selectCursor(type, filter)
+        error("游标查询必须在 transaction {} 中使用")
     }
 
     override fun sortCursor(row: String, filter: Filter.() -> Unit): Cursor<T> {
-        return operator.sortCursor(type, row, filter)
+        error("游标查询必须在 transaction {} 中使用")
     }
 
     override fun sortDescendingCursor(row: String, filter: Filter.() -> Unit): Cursor<T> {
-        return operator.sortDescendingCursor(type, row, filter)
+        error("游标查询必须在 transaction {} 中使用")
     }
 
     // === 更新 ===
@@ -176,17 +210,18 @@ class TransactionalDataMapper<T>(
 
     // === 检查 ===
 
-    override fun exists(id: Any, filter: Filter.() -> Unit): Boolean = operator.has(type, id, filter)
-    override fun exists(filter: Filter.() -> Unit): Boolean = operator.has(filter)
+    override fun exists(id: Any, filter: Filter.() -> Unit): Boolean {
+        return executeBeanRead("exists", id, filter) { operator.has(type, id, filter) }
+    }
+
+    override fun exists(filter: Filter.() -> Unit): Boolean {
+        return executeRead("existsFilter", filter) { operator.has(filter) }
+    }
 
     // === 计数 ===
 
-    override fun count(filter: Filter.() -> Unit): Long = operator.count(filter)
-
-    // === 事务 ===
-
-    override fun <R> transaction(block: DataMapper<T>.() -> R): Result<R> {
-        error("Nested transactions are not supported")
+    override fun count(filter: Filter.() -> Unit): Long {
+        return executeRead("count", filter) { operator.count(filter) }
     }
 
     // === 自定义 SQL ===
@@ -195,20 +230,20 @@ class TransactionalDataMapper<T>(
         get() = operator.table.name
 
     override fun query(builder: ActionSelect.() -> Unit): List<T> {
-        val typeClass = AnalyzedClass.of(type)
+        val mapper = EntityMapper.of(type)
         val action = ActionSelect(tableName).apply(builder)
         return operator.select(action) { rs ->
             buildList {
-                while (rs.next()) { add(typeClass.createInstance<T>(typeClass.read(rs))) }
+                while (rs.next()) { add(mapper.createInstance(mapper.read(rs))) }
             }
         }
     }
 
     override fun queryOne(builder: ActionSelect.() -> Unit): T? {
-        val typeClass = AnalyzedClass.of(type)
+        val mapper = EntityMapper.of(type)
         val action = ActionSelect(tableName).apply { builder(); limit(1) }
         return operator.select(action) { rs ->
-            if (rs.next()) typeClass.createInstance<T>(typeClass.read(rs)) else null
+            if (rs.next()) mapper.createInstance(mapper.read(rs)) else null
         }
     }
 
@@ -237,21 +272,6 @@ class TransactionalDataMapper<T>(
         return result
     }
 
-    // === 多表联查 ===
-
-    override fun join(builder: JoinQuery.() -> Unit): JoinQuery {
-        return JoinQuery(operator.dataSource, sharedConnection).also {
-            it.from(tableName)
-            builder(it)
-        }
-    }
-
-    // === 生命周期 ===
-
-    override fun close() {
-        error("Cannot close container within a transaction")
-    }
-
     // === 容器类型 Accessor ===
 
     override fun mapOf(id: Any, fieldName: String): MutableMap<String, String?> {
@@ -278,28 +298,44 @@ class TransactionalDataMapper<T>(
         return (operator as ContainerOperatorImpl).setAccessor(fieldName, filter)
     }
 
-    // === L2 缓存失效策略 ===
+    // === 缓存辅助 ===
 
-    private fun buildCacheKey(method: String, vararg args: Any?): String {
+    @Suppress("UNCHECKED_CAST")
+    protected fun <R> cachedBean(method: String, vararg args: Any?, query: () -> R): R {
+        val c = cache ?: return query()
+        val key = buildCacheKey(method, *args)
+        return c.beanCache.get(key) { query() } as R
+    }
+
+    @Suppress("UNCHECKED_CAST")
+    protected fun <R> cachedQuery(method: String, vararg args: Any?, query: () -> R): R {
+        val c = cache ?: return query()
+        val key = buildCacheKey(method, *args)
+        return c.queryCache.get(key) { query() } as R
+    }
+
+    protected fun buildCacheKey(method: String, vararg args: Any?): String {
         return "$method:${args.joinToString(",") { it?.toString() ?: "null" }}"
     }
 
+    // === L2 缓存失效策略 ===
+
     /** 插入后：仅清空 Query Cache */
-    private fun invalidateOnInsert() {
+    protected fun invalidateOnInsert() {
         cache?.queryCache?.invalidateAll()
     }
 
     /** 单条更新/删除后（按 ID）：失效该 ID 的 Bean Cache + 清空 Query Cache */
-    private fun invalidateOnIdMutation(id: Any) {
-        if (cache == null) return
-        cache.beanCache.invalidateByPrefix("findById:$id")
-        cache.beanCache.invalidateByPrefix("findAllById:$id")
-        cache.beanCache.invalidateByPrefix("exists:$id")
-        cache.queryCache.invalidateAll()
+    protected fun invalidateOnIdMutation(id: Any) {
+        val c = cache ?: return
+        c.beanCache.invalidateByPrefix("findById:$id")
+        c.beanCache.invalidateByPrefix("findAllById:$id")
+        c.beanCache.invalidateByPrefix("exists:$id")
+        c.queryCache.invalidateAll()
     }
 
     /** 从数据对象提取 @Id 值，失效该 ID 的 Bean Cache + 清空 Query Cache */
-    private fun invalidateOnMutation(data: Any) {
+    protected fun invalidateOnMutation(data: Any) {
         if (cache == null) return
         val id = analyzedClass.getPrimaryMemberValue(data)
         if (id != null) {
@@ -310,14 +346,14 @@ class TransactionalDataMapper<T>(
     }
 
     /** 失效指定行 ID 的 Bean Cache + 清空 Query Cache */
-    private fun invalidateOnRowIdMutation(rowId: Long) {
-        if (cache == null) return
-        cache.beanCache.invalidate(buildCacheKey("findByRowId", rowId))
-        cache.queryCache.invalidateAll()
+    protected fun invalidateOnRowIdMutation(rowId: Long) {
+        val c = cache ?: return
+        c.beanCache.invalidate(buildCacheKey("findByRowId", rowId))
+        c.queryCache.invalidateAll()
     }
 
     /** 批量/不确定范围：全部清空 */
-    private fun invalidateAll() {
+    protected fun invalidateAll() {
         cache?.invalidateAll()
     }
 }
