+++ module/database/database-ptc-object/src/test/kotlin/taboolib/expansion/ContainerOperatorTest.kt
@@ -5,6 +5,8 @@ import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
+import taboolib.expansion.operator.ContainerOperatorImpl
+import taboolib.expansion.orm.AnalyzedClass
 
 class ContainerOperatorTest {
