+++ module/database/database-ptc-object/src/test/kotlin/taboolib/expansion/JoinQueryTest.kt
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
+import taboolib.expansion.orm.AnalyzedClass
 
 /**
  * JoinQuery 多表联查测试
