+++ module/database/database-ptc-object/src/test/kotlin/taboolib/expansion/TransactionPropagationTest.kt
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
+import taboolib.expansion.orm.AnalyzedClass
 
 class TransactionPropagationTest {
